+++ packages/config-provider/index.ts
@@ -1,0 +1,10 @@
+import { VantComponent } from '../common/component';
+
+VantComponent({
+  props: {
+    themeVars: {
+      type: Object,
+      value: {},
+    },
+  },
+});