+++ packages/grid-item/index.ts
@@ -12,6 +12,10 @@ VantComponent({
   props: {
     icon: String,
     iconColor: String,
+    iconPrefix: {
+      type: String,
+      value: 'van-icon',
+    },
     dot: Boolean,
     info: null,
     badge: null,
@@ -42,6 +46,7 @@ VantComponent({
         clickable,
         center,
         direction,
+        reverse,
         iconSize,
       } = data;
 
@@ -52,6 +57,7 @@ VantComponent({
         gutter,
         clickable,
         direction,
+        reverse,
         iconSize,
         index: children.indexOf(this),
         columnNum,