+++ docs/markdown/changelog.md
@@ -1,5 +1,84 @@
 # 更新日志
 
+
+### [1.7.2](https://github.com/youzan/vant-weapp/tree/v1.7.2)
+
+`2021-07-19`
+
+**Bug Fixes**
+
+ - Calendar: 初始日期设置为当前日期 [#4339](https://github.com/youzan/vant-weapp/issues/4339)
+
+**Features**
+
+ - Cell: CellGroup 新增 inset 属性 [#4341](https://github.com/youzan/vant-weapp/issues/4341)
+ - Search: 新增click-input 事件 [#4337](https://github.com/youzan/vant-weapp/issues/4337)
+
+### [1.7.1](https://github.com/youzan/vant-weapp/tree/v1.7.1)
+
+`2021-07-06`
+
+**Bug Fixes**
+
+- Col: 修复样式问题 [#4322](https://github.com/youzan/vant-weapp/issues/4322)
+- Grid: 增加 icon-prefix 属性默认值 [#4318](https://github.com/youzan/vant-weapp/issues/4318)
+- IndexBar: 修复当索引列表为空时报错 [#4310](https://github.com/youzan/vant-weapp/issues/4310)
+
+### [1.7.0](https://github.com/youzan/vant-weapp/tree/v1.7.0)
+
+`2021-07-04`
+
+**Bug Fixes**
+
+- Button: 修复 type 为 getUserInfo & wx.getUserProfile 可用时 lang 参数失效 [#4250](https://github.com/youzan/vant-weapp/issues/4250)
+- ShareSheet: 新增内置icon `weapp-qrcode` `wechat-moments` [#4256](https://github.com/youzan/vant-weapp/issues/4256)
+
+**Features**
+
+- Checkbox: 增加 direction 属性，设置排列方向 [#4265](https://github.com/youzan/vant-weapp/issues/4265)
+- ConfigProvider: 增加全局配置组件 [#4279](https://github.com/youzan/vant-weapp/issues/4279)
+- Grid: 增加 reverse 属性，支持文本&图片位置互换 [#4280](https://github.com/youzan/vant-weapp/issues/4280)
+- GridItem: 增加 icon-prefix 属性，支持第三方字体 [#4276](https://github.com/youzan/vant-weapp/issues/4276)
+- Rate: 增加更多 css 变量 [#4297](https://github.com/youzan/vant-weapp/issues/4297)
+- Slider: 增加更多 css 变量 [#4305](https://github.com/youzan/vant-weapp/issues/4305)
+
+**Improvements**
+
+- wxs style 方法支持驼峰变量命名，以便能更好的在wxml中使用 [#4281](https://github.com/youzan/vant-weapp/issues/4281)
+
+### [1.6.9](https://github.com/youzan/vant-weapp/tree/v1.6.9)
+
+`2021-06-06`
+
+**Bug Fixes**
+
+- DatetimePicker: 修复 动态设置 min-hour min-date 显示不正确 [#4245](https://github.com/youzan/vant-weapp/issues/4245)
+- Tabs: 修复 tab数量较多时滚动距离不正确 [#4202](https://github.com/youzan/vant-weapp/issues/4202)
+- Uploader: 修复 demo 中 beforeRead 报错 [#4235](https://github.com/youzan/vant-weapp/issues/4235)
+
+### [1.6.9-beta.2](https://github.com/youzan/vant-weapp/tree/v1.6.9-beta.2)
+
+`2021-05-22`
+
+**Bug Fixes**
+
+- Button: 修复 open-type 不生效 [#4222](https://github.com/youzan/vant-weapp/issues/4222)
+- Calendar: 修复 allow-same-day 开启时 custom-color 属性无效 [#4200](https://github.com/youzan/vant-weapp/issues/4200)
+- Circle: 修复 value 是小数时抖动的情况 [#4152](https://github.com/youzan/vant-weapp/issues/4152)
+- NoticeBar: 调整滚动行为 [#4201](https://github.com/youzan/vant-weapp/issues/4201)
+- NoticeBar: 修复不正确的滚动速度 [cde3876](https://github.com/youzan/vant-weapp/commit/cde3876fb0742cacf3e481a8eb2b487dabc8709e)
+- Radio: 支持动态设置 disabled [#4191](https://github.com/youzan/vant-weapp/issues/4191)
+- Rate: 支持滑动选择半星 [#4195](https://github.com/youzan/vant-weapp/issues/4195)
+
+**Features**
+
+- Calendar: 增加 first-day-of-week 属性，设置周起始日 [#4211](https://github.com/youzan/vant-weapp/issues/4211)
+- Calendar: 增加 show-range-prompt 属性和 over-range 事件，设置是否展示提示文案 [#4212](https://github.com/youzan/vant-weapp/issues/4212)
+- Icons: 图标库迁移至iconfont.cn [#4219](https://github.com/youzan/vant-weapp/issues/4219)
+- OpenType: 支持 getUserProfile [#4203](https://github.com/youzan/vant-weapp/issues/4203)
+- Panel: 移除 useFooterSlot [#4205](https://github.com/youzan/vant-weapp/issues/4205)
+
+
 ### [1.6.8](https://github.com/youzan/vant-weapp/tree/v1.6.8)
 
 `2021-02-26`